Fix use of incorrect NCCL error code
ncclInvalidUsage is not an error code that should be returned by the
plugin.  ncclInvalidArgument is the plugin-approved error code, so
move to ncclInvalidArgument.

ncclInvalidUsage is not in the list of error codes plugins can call and
nvidida's new compat headers removes it, breaking our build when
we import the new headers.

Signed-off-by: Brian Barrett <bbarrett@amazon.com>