Add inline_write method for Neuron.
Neuron needs the ability to perform a local write to a local address via
an EFA device.  To add support, fi_inject write is called via the
inline_write method in the send or receive comm objects.  The write
input data is immediately released, and the write completes, but does
not provide a completion to the caller.