Separate the domain, and abstract access to the domain.
Create two paths for creating the domain: per endpoint or per device.
This is because, on some platforms, we lock the domain when accessing the
endpoint, so retaining separate domains per thread and per endpoint reduces contention.  This
depends on the threading model used with libfabric.

Create an accessor method, get_domain(), that abstractly accesses the endpoint by protocol
and architecture.

Signed-off-by: Ryan Hankins <rqh@amazon.com>